resolved PR comments and other bug fixes

+++ src/commands/queryHistory/deleteQuery.ts
@@ -4,7 +4,13 @@ import { deleteQuery, getQueryHistory } from "../../util/queryHistory";
 import { QueryHistoryTreeProvider } from '../../tree/QueryHistoryTreeProvider';
 
 export const deleteQueryItem = async (delQuery: IQuery, queryHistoryTreeProvider: QueryHistoryTreeProvider) => {
-    deleteQuery(delQuery);
-    queryHistoryTreeProvider.refresh();
-    vscode.window.showInformationMessage("Query deleted successfully from history");
+    vscode.window.showWarningMessage(`Are you sure you want to delete the query from query history? Query: ${delQuery.query}`,
+    "Yes",
+    "No").then((value)=>{
+        if(value === "Yes"){
+            deleteQuery(delQuery);
+            queryHistoryTreeProvider.refresh();
+            vscode.window.showInformationMessage("Query deleted successfully from history");
+        }
+    });
 };
