+++ src/pages/FavoriteQueries/FavoriteQueries.ts
@@ -6,6 +6,7 @@ import { applyQuery } from '../../commands/queryHistory/applyQuery';
 import { Memory, getUUID } from '../../util/util';
 import { deleteFavoriteQuery } from '../../util/favoriteQuery';
 import { IConnection } from '../../types/IConnection';
+import { Constants } from '../../util/constants';
 
 export interface IFavoriteQueriesWebviewParams {
     styleSrc: vscode.Uri;
@@ -16,16 +17,11 @@ export interface IFavoriteQueriesWebviewState {
 }
 
 export const fetchFavoriteQueries = (context: vscode.ExtensionContext) => {
-    const connection = Memory.state.get<IConnection>("activeConnection");
-    if (!connection) {
-        vscode.window.showErrorMessage("Please connect to a cluster before opening favorite queries");
-        return;
-    }
-    const favoriteQueryWebviewDetails = Memory.state.get<IFavoriteQueriesWebviewState>("favoriteQueriesWebview");
+    const favoriteQueryWebviewDetails = Memory.state.get<IFavoriteQueriesWebviewState>(Constants.FAVORITE_QUERIES_WEBVIEW);
     if (favoriteQueryWebviewDetails) {
         // Favorite Queries Webview already exists, Closing existing and creating new
         favoriteQueryWebviewDetails.webviewPanel.dispose();
-        Memory.state.update("favoriteQueriesWebview", null);
+        Memory.state.update(Constants.FAVORITE_QUERIES_WEBVIEW, null);
     }
     const currentPanel = vscode.window.createWebviewPanel(
         "showFavoriteQueries",
@@ -40,8 +36,6 @@ export const fetchFavoriteQueries = (context: vscode.ExtensionContext) => {
     try {
         const onDiskPath = vscode.Uri.file(path.join(context.extensionPath, 'src/webviews/styles/favoriteQueries.css'));
         const styleSrc = currentPanel.webview.asWebviewUri(onDiskPath);
-        const onDiskPathEditLogo = vscode.Uri.joinPath(context.extensionUri, 'images', 'edit-icon.svg');
-        const editLogo = currentPanel.webview.asWebviewUri(onDiskPathEditLogo);
         const UriData: IFavoriteQueriesWebviewParams = {
             styleSrc: styleSrc
         };
@@ -49,8 +43,12 @@ export const fetchFavoriteQueries = (context: vscode.ExtensionContext) => {
         currentPanel.webview.onDidReceiveMessage(async (message) => {
             switch (message.command) {
                 case 'vscode-couchbase.pasteQuery':
-                    // Handle the message received from the webview
                     const dataFromWebview = message.query;
+                    const connection = Memory.state.get<IConnection>(Constants.ACTIVE_CONNECTION);
+                    if (!connection) {
+                        vscode.window.showErrorMessage("Please connect to a cluster before pasting a query");
+                        return;
+                    }
                     currentPanel.dispose();
                     await new Promise((resolve) => setTimeout(resolve, 200));
                     applyQuery({ query: dataFromWebview, id: getUUID() });
@@ -58,23 +56,31 @@ export const fetchFavoriteQueries = (context: vscode.ExtensionContext) => {
                     break;
                 case 'vscode-couchbase.deleteQuery':
                     const queryId = message.id;
-                    deleteFavoriteQuery(queryId, context);
+                    vscode.window.showWarningMessage(`Are you sure you want to delete the query from favorite queries? Query: ${message.query}`,
+                        "Yes",
+                        "No").then((value)=>{
+                            if(value === "Yes"){
+                                deleteFavoriteQuery(queryId, context);
+                            }
+                        });
                     break;
-                // Add more cases for other message types if needed
+                case 'vscode-couchbase.queryNotSelected':
+                    vscode.window.showErrorMessage("Please select a query before performing an action");
+                    break;
             }
         });
 
         currentPanel.onDidDispose(() => {
-            Memory.state.update("favoriteQueriesWebview", null);
+            Memory.state.update(Constants.FAVORITE_QUERIES_WEBVIEW, null);
         });
 
-        Memory.state.update("favoriteQueriesWebview", {
+        Memory.state.update(Constants.FAVORITE_QUERIES_WEBVIEW, {
             webviewPanel: currentPanel
         });
 
 
     } catch (err) {
-        logger.error("failed to open and set query context: " + err);
+        logger.error(`failed to open and set query context: ${err}`);
         logger.debug(err);
     }
 };
